WiP: now having a full functional migration of refSO3 (with positive and negatif pfn_offset)

+++ agency/usr/include/core/core.h
@@ -47,7 +47,7 @@ int get_ME_free_slot(size_t ME_size);
 
 bool get_ME_id(uint32_t slotID, ME_id_t *ME_id);
 
-void read_ME_snapshot(unsigned int slotID, void **buffer, size_t *buffer_size);
+void read_ME_snapshot(unsigned int slotID, void **buffer, uint32_t *buffer_size);
 void write_ME_snapshot(unsigned int slotID, unsigned char *ME_buffer);
 
 void *prepare_ME_slot(unsigned int slotID);
